Remove 'disposition' column from 'calls' table migration

+++ database/migrations/2024_02_19_115726_remove_disposition_column_from_call_table.php
@@ -11,11 +11,13 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('calls', function (Blueprint $table) {
-            if (Schema::hasColumn('calls', 'disposition')) { // Check if the column exists
-                $table->dropColumn('disposition');
-            }
-        });
+        if (Schema::hasTable('calls')) {
+            Schema::table('calls', function (Blueprint $table) {
+                if (Schema::hasColumn('calls', 'disposition')) { // Check if the column exists
+                    $table->dropColumn('disposition');
+                }
+            });
+        }
     }
 
     /**
@@ -23,10 +25,12 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('calls', function (Blueprint $table) {
-            if (!Schema::hasColumn('calls', 'disposition')) { // Check if the column doesn't exist
-                $table->string('disposition')->nullable()->after('call_type_id');
-            }
-        });
+        if (Schema::hasTable('calls')) {
+            Schema::table('calls', function (Blueprint $table) {
+                if (!Schema::hasColumn('calls', 'disposition')) { // Check if the column doesn't exist
+                    $table->string('disposition')->nullable()->after('call_type_id');
+                }
+            });
+        }
     }
 };
